[refactor] 메세지 상수로 변경, 폴더 변경 #158
